+++ example/http/server/server.go
@@ -23,8 +23,8 @@ import (
 	"github.com/open-telemetry/opentelemetry-go/api/trace"
 	"github.com/open-telemetry/opentelemetry-go/plugin/httptrace"
 
-	"github.com/open-telemetry/opentelemetry-go/api/event"
 	_ "github.com/open-telemetry/opentelemetry-go/exporter/loader"
+	"github.com/open-telemetry/opentelemetry-go/sdk/event"
 )
 
 var (
