+++ plugin/httptrace/clienttrace.go
@@ -22,9 +22,10 @@ import (
 	"strings"
 
 	"github.com/open-telemetry/opentelemetry-go/api/core"
-	"github.com/open-telemetry/opentelemetry-go/api/event"
 	"github.com/open-telemetry/opentelemetry-go/api/tag"
 	"github.com/open-telemetry/opentelemetry-go/api/trace"
+
+	"github.com/open-telemetry/opentelemetry-go/sdk/event"
 )
 
 type (
